feat: restructure search panel classes
